+++ openquake/hazardlib/calc/conditioned_gmfs.py
@@ -114,8 +114,9 @@ import psutil
 import numpy
 import pandas
 from openquake.baselib import parallel
+from openquake.hazardlib.truncated_mvn import TruncatedMVN
 from openquake.hazardlib import correlation, cross_correlation
-from openquake.hazardlib.calc.gmf import GmfComputer, TRUNCATION_THRESHOLD
+from openquake.hazardlib.calc.gmf import exp, GmfComputer, TRUNCATION_THRESHOLD
 from openquake.hazardlib.const import StdDev
 from openquake.hazardlib.geo.geodetic import geodetic_distance
 
@@ -229,6 +230,51 @@ class ConditionedGmfComputer(GmfComputer):
             spatial_correl or correlation.JB2009CorrelationModel(clust),
             cross_correl_between or cross_correlation.GodaAtkinson2009(),
             cross_correlation.BakerJayaram2008())
+
+    def _compute(self, muWWBB, m, imt, gsim, within_eps, idxs):
+        # mea, tau, phi with shapes (N, 1), (N, N), (N, N)
+        mu_Y, cov_WY_WY, cov_BY_BY = muWWBB
+        E = len(idxs)
+        if max(self.tlw, self.tlb) <= TRUNCATION_THRESHOLD:
+            arr = mu_Y.repeat(E, axis=1)
+        else:
+            arr = self._compute_mvn(cov_WY_WY, cov_BY_BY, mu_Y, E)
+        gmf = exp(arr, imt != "MMI")
+        return gmf  # shapes (N, E)
+
+    def _compute_mvn(self, cov_WY_WY, cov_BY_BY, mu_Y, E):
+        N = len(cov_WY_WY)
+        cutoff = numpy.eye(N) * self.cmaker.oq.correlation_cutoff
+        # the cutoff is needed to remove negative eigenvalues
+        if (self.cmaker.oq.truncated_mvn is False or
+                self.cmaker.truncation_level == 99):
+            # do not truncate
+            cov_Y_Y = cov_WY_WY + cov_BY_BY + cutoff
+            arr = self.rng.multivariate_normal(
+                mu_Y.flatten(), cov_Y_Y, size=E,
+                check_valid="raise", tol=1e-5, method="cholesky")
+            return arr.T
+
+        # NB: truncated MVN is used in the scenario risk tests
+        # conditioned_stations, case_21_stations, case_26_stations
+        cov_WY_WY = cov_WY_WY + cutoff
+        cov_BY_BY = cov_BY_BY + cutoff
+
+        lb_w, ub_w = self.get_symmetric_bounds(cov_WY_WY, self.tlw)
+        seed_w = int(self.rng.integers(0, numpy.iinfo(numpy.int32).max))
+
+        z_w_truncated = TruncatedMVN(
+            numpy.zeros(N), cov_WY_WY, lb_w, ub_w, seed=seed_w
+        ).sample(E)
+
+        lb_b, ub_b = self.get_symmetric_bounds(cov_BY_BY, self.tlb)
+        seed_b = int(self.rng.integers(0, numpy.iinfo(numpy.int32).max))
+        z_b_truncated = TruncatedMVN(
+            numpy.zeros(N), cov_BY_BY, lb_b, ub_b, seed=seed_b
+        ).sample(E)
+
+        arr = mu_Y.flatten()[:, None] + z_w_truncated + z_b_truncated
+        return arr
 
     # parallelized
     def get_mea_tau_phi(self, h5):
